pyproject.toml
[project]
name = "podcast-langgraph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12 "
dependencies = [
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.68",
    "langchain-openai>=0.3.27",
    "langchain>=0.3.26",
    "langgraph>=0.5.1",
    "pillow>=11.3.0",
    "pydub>=0.25.1",
    "simpleaudio>=1.0.4",
    "wikipedia>=1.4.0",
]
